feat: 전역 대화 상자들 상태 관리를 위한 AppProviders 구성, App 컴포넌트 app 계층으로 이동
